+++ qipfs/go-ipfs-files/file_test.go
