+++ qipfs/go-ipfs-files/is_hidden.go
