+++ qipfs/go-ipfs-files/is_hidden_windows.go
