+++ qipfs/go-ipfs-files/linkfile.go
