+++ qipfs/go-ipfs-files/multifilereader.go
