+++ qipfs/go-ipfs-files/multifilereader_test.go
