+++ qipfs/go-ipfs-files/readerfile.go
