+++ qipfs/go-ipfs-files/slicefile.go
