+++ qipfs/go-ipfs-files/webfile_test.go
